Cria a regra de exclusão de pedidos

+++ lambda/orders/ordersFunction.ts
@@ -32,15 +32,15 @@ export async function handler(event: APIGatewayProxyEvent, context: Context): Pr
     const lambdaRequestId = context.awsRequestId
 
     console.log(`API Gateway RequestId: ${apiRequestId} - LambdaRequestId: ${lambdaRequestId}`)
-
+    
+    const isAdmin = authInfoService.isAdminUser(event.requestContext.authorizer)
+    const authenticatedUser = await authInfoService.getUserInfo(event.requestContext.authorizer)
+    
     if(method === 'GET'){
         if(event.queryStringParameters){
             const email = event.queryStringParameters!.email
             const orderId = event.queryStringParameters!.orderId
 
-            const isAdmin = authInfoService.isAdminUser(event.requestContext.authorizer)
-            const authenticatedUser = await authInfoService.getUserInfo(event.requestContext.authorizer)
-            
             if(isAdmin || email === authenticatedUser){
                 
                 if(email){
@@ -73,7 +73,7 @@ export async function handler(event: APIGatewayProxyEvent, context: Context): Pr
                 }
             }
         } else {
-            if(authInfoService.isAdminUser(event.requestContext.authorizer)){
+            if(isAdmin){
                 const orders = await orderRepository.getAllOrders()
                 return {
                     statusCode: 200,
@@ -135,24 +135,31 @@ export async function handler(event: APIGatewayProxyEvent, context: Context): Pr
         const email = event.queryStringParameters!.email!
         const orderId = event.queryStringParameters!.orderId!
 
-        try{
-            const orderDelete = await orderRepository.deleteOrder(email, orderId)
-            const eventResult = await sendOrderEvent(orderDelete, OrderEventType.DELETED, lambdaRequestId)
-            console.log (
-                `Order deleted event sent - OrderId: ${orderDelete.sk} - MessageId: ${eventResult.MessageId}`
-            )
-
-            return {
-                statusCode: 200,
-                body: JSON.stringify(convertToOrderResponse(orderDelete))
+        if(isAdmin || email === authenticatedUser){
+            try{
+                const orderDelete = await orderRepository.deleteOrder(email, orderId)
+                const eventResult = await sendOrderEvent(orderDelete, OrderEventType.DELETED, lambdaRequestId)
+                console.log (
+                    `Order deleted event sent - OrderId: ${orderDelete.sk} - MessageId: ${eventResult.MessageId}`
+                )
+    
+                return {
+                    statusCode: 200,
+                    body: JSON.stringify(convertToOrderResponse(orderDelete))
+                }
+            } catch (error){
+                console.log((<Error>error).message)
+                return {
+                    statusCode: 404,
+                    body: (<Error>error).message
+                }
             }
-        } catch (error){
-            console.log((<Error>error).message)
+        } else {
             return {
-                statusCode: 404,
-                body: (<Error>error).message
+                statusCode: 403,
+                body: `You don't have permission to access this operation`
             }
-        }
+        }        
     }
 
     return {
